Allow f32 and f64 to co-exist
