+++ src/common/steps/broad_phase.rs
@@ -1,4 +1,4 @@
-use crate::prelude::*;
+use super::super::prelude::*;
 use bevy::{prelude::*, utils::StableHashSet};
 
 pub struct BroadPhasePlugin;
