+++ src/common/steps/integrator.rs
@@ -1,4 +1,4 @@
-use crate::prelude::*;
+use super::super::prelude::*;
 use bevy::prelude::*;
 
 pub struct IntegratorPlugin;
